Refactored movie reviews into its own package, minor layout changes for reviews
